♻️refactor: ConversationCompleteCommandService 분리
